Improve model & implement self-play.
Add temporary workaround so that all players are symmetrically equivalent.
Change network archiecture so that previous choices influence subsequent choices. For example, when choosing which card to use, that part of the model knows that maybe we've decided to do a discard action.
Add functionality that allows us to instantiate an inference class by copying the models from the training class.